requirements.txt
PyGithub
gitpython
matplotlib
javalang
plyj
plantuml
